chimera-installer: add
This is a working skeleton of a Chimera installer. There have been
a few attempts at writing one over the previous weeks, first trying
a Python version, but all attempts proved to be problematic in some
way.

In the end I decided to write one in shell and bundle it with the
install scripts. This version is also following a few iterations.

There was code for filesystem validation and mounting, but I was
not happy with it yet so I ripped it out; the previous version did
validation based on the mount table as set up by the UI, but this
is clunky when it comes to integration with custom preexisting
mounts, so I will write a new version in the coming days, which
will create filesystems without validation, mount them (on top of
any existing structure) and then perform validation based on
runtime state.

I also have some WIP code for handling bootloader installation.
I am not happy with it yet so I have not included it.

All these things will be written in time for a new set of images
during the coming week; I have local prototypes but not yet happy
with them.